pyproject.toml
[project]
name = "ScanServer"
version = "0.1.0"
description = "Local COM scanner client for CustomCraft MES"
requires-python = ">=3.12,<3.13"
dependencies = [
  "pyserial>=3.5",
  "httpx>=0.27.0",
  "edge-tts>=6.1.10",
  "pygame>=2.5.2",
  "pyttsx3>=2.90",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
]

[tool.ruff]
line-length = 110
target-version = "py312"
